Move code generation to `msg-gen` crate
